requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
cymem==2.0.3
Django==2.2.14
djangorestframework==3.11.0
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
idna==2.10
importlib-metadata==1.7.0
joblib==0.14.1
jsonschema==3.2.0
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
pandas==0.24.2
pdfminer.six==20200726
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
pycparser==2.20
pycryptodome==3.9.8
pyresparser==1.0.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.1
tqdm==4.48.0
urllib3==1.25.10
wasabi==0.7.1
zipp==1.2.0
